Added fileSize to Envelope + EnvelopeRepo

+++ src/main/java/uk/gov/hmcts/reform/blobrouter/data/envelopes/Envelope.java
@@ -14,7 +14,7 @@ public class Envelope {
     public final Status status;
     public final boolean isDeleted;
     public final boolean pendingNotification;
-    public final int fileSize;
+    public final Long fileSize;
 
     public Envelope(
         UUID id,
@@ -26,7 +26,7 @@ public class Envelope {
         Status status,
         boolean isDeleted,
         boolean pendingNotification,
-        int fileSize
+        Long fileSize
     ) {
         this.id = id;
         this.container = container;
